Split out Example component and use that with ButtonExample
