+++ aries-site/src/examples/components/controls/ButtonExample.js
@@ -2,23 +2,9 @@ import React from 'react';
 import { Box, Button } from 'grommet';
 import { Next } from 'grommet-icons';
 
-import { UsageExample } from '../../../layouts';
-
 export const ButtonExample = () => {
   return (
-    <UsageExample
-      pad={{
-        vertical: 'large',
-        horizontal: 'large',
-        small: { horizontal: 'large', top: 'medium', bottom: 'xlarge' },
-      }}
-      direction="column"
-      align="start"
-      gap="large"
-      docs="https://v2.grommet.io/button?theme=hpe"
-      code="https://raw.githubusercontent.com/hpe-design/design-system/master/aries-site/src/examples/components/controls/MenuExample.js"
-      figma="https://www.figma.com/file/HZq1i3avh8VLK4kWXs38sW/grommet-hpe-form?node-id=0%3A1"
-    >
+    <Box direction="column" align="start" gap="large">
       <Button
         label="Check it out!"
         icon={<Next />}
@@ -31,6 +17,6 @@ export const ButtonExample = () => {
         <Button type="reset" label="Reset" disabled onClick={() => {}} />
         <Button label="Cancel" onClick={() => {}} />
       </Box>
-    </UsageExample>
+    </Box>
   );
 };
