+++ aries-site/src/layouts/content/index.js
@@ -1,6 +1,7 @@
 export * from './ButtonRow';
 export * from './ColorRow';
 export * from './ContentSection';
+export * from './Example';
 export * from './Head';
 export * from './Subsection';
 export * from './TypographyRow';
