+++ aries-site/src/pages/_app.js
@@ -1,3 +1,4 @@
+// import '../../../node_modules/prismjs/themes/prism.css';
 import '../../../node_modules/prismjs/themes/prism-tomorrow.css';
 
 // This default export is required in a new `pages/_app.js` file.
